fixed compilation issues for macos on gcc. clang still yet to be fixed

+++ neurox/neurox_hpx.h
@@ -14,7 +14,7 @@
 typedef hpx_addr_t hpx_t;    ///> hpx address (just rephrased with shorter naming)
 typedef double floble_t;     ///> float or double (v, matrix values and mechanisms)
 typedef int offset_t;        ///> ushort or uint (p vector, nodes indices)
-typedef int64_t neuron_id_t; ///> neuron gids (gid_t or id_t already used by types.h)
+typedef int neuron_id_t; ///> neuron gids (gid_t or id_t already used by types.h)
                               // (neuron id is past as -1 on input params parcelgid argument)
 
 //Memory alignment for hpx_gas_allocs
